Share Synapse between dev environments
The Synapse SQL DataWarehouse is super expensive compared to the rest of
the infrastructure. This moves Synapse to the shared dev environment,
which means we end up with a single instance in dev and a single
instance in production.

We can revisit this later on and potentially move it back to the
individual environments when the need arises.