merge doesn't work on shallow clones
This is annoying, because you'd think that git would be
able to work out that I want to pull, and therefore it
should deepen its clones of both branches until they
either have a common ancestor or are proven to really
not share a root. Maybe it's not that simple.