Disable certain actions if secrets are not available
Fork PRs don't have access to secrets. The most reliable way to check for this
seems to be checking the secret values.

Here we just skip steps, that require secrets. Those steps are pushing Docker
images to our Azure registry and pushing updated manifests to the depployments
repo. These checks make it easy to test the changes on a dev environment, but
they aren't otherwise necessary. So it should be okay to skip them.