Correlate GraphQL traces with request
The tracing extension, that comes with async-graphql currently forces
every GraphQL query to be in a new trace. We want it to be part of the
request trace, though. This vendors the extension. I will suggest a
change the async-graphql later and see if we can get this upstreamed.

I also noticed async-graphql doesn't call the error handler of the
extension anymore since version 2.0.9. This is why I fixed the version
to 2.0.8 for now. Similarly I will open an issue for this.

This commit also fixes an issue with the B3Propagator, which is
explained in a code comment.