Fix yarn: Extracting tar content of undefined failed
We recently added nhsuk-react-components as a git dependency, because we
want to use features from the latest master branch, which have not been
released, yet.

This has caused a bug, where builds were randomly failing with the
error:

    yarn install v1.22.5
    [1/4] Resolving packages...
    [2/4] Fetching packages...
    [1/4] Resolving packages...
    [2/4] Fetching packages...
    error https://registry.yarnpkg.com/yaml/-/yaml-1.10.0.tgz: Extracting tar content of undefined failed, the file appears to be corrupt: "ENOENT: no such file or directory, chmod '/home/runner/.cache/yarn/v6/npm-yaml-1.10.0-3b593add944876077d4d683fee01081bd9fff31e-integrity/node_modules/yaml/util.mjs'"

It looks like this issue: https://github.com/yarnpkg/yarn/issues/7212

This fixes it by publishing a version of the nhsuk-react-components
package to a GitHub npm registry under our organization. Our frontend
then fetches the package from there.